Adds responsive carousel to case studies and other improvements
Integrates a responsive carousel component into the Case Studies section for a more user-friendly experience.
Improves image display and captioning in the Case Studies component.
Updates the images used in the Case Studies section.
Refactors the BlogSection component, adding a button for more information.
Modifies the appearance of the map and modal components for a more polished user interface.
Changes the order of components in the MainSection.
Updates the navbar with social media icons using React Icons.